vincular-notion/requirements.txt
# Versão: 1.0
# Dependências do projeto

notion-client==2.2.1
python-dotenv==1.0.0
